Add support for semihosting: WITH_SEMIHOSTING=1
Semihosting is a fun feature that let's a microcontroller use a couple of
syscalls on its host (the computer being connected to the micro through JTAG).

This includes stuff like outputting data, as if the micro had a terminal con-
nected to it as well as doing file IO on the host computer.

This uses "printf()" to print some data to the host's terminal. I've tested
this using OpenOCD 0.10.0. You may need to issue

  monitor arm semihosting on

from a gdb prompt connected to the OpenOCD server.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>